+++ test/helper.rb
@@ -1,2 +1,5 @@
 require 'minitest/autorun'
 require 'minitest/pride'
+
+require_relative 'support/test_enumerable_source'
+require_relative 'support/assert_called'