+++ Changes.md
@@ -1,2 +1,5 @@
 HEAD
 ----
+
+- New: Kiba::Common::DSLExtensions::ShowMe (useful to color-print rows during development)
+- New: Kiba::Common::DSLExtensions::Logger (production logging)